Perf - Move config file check into backends
